vqa_generation/requirements.txt
anthropic==0.57.1
arxiv==2.1.0
datasets==2.16.1
deepsearch==0.1
fitz==0.0.1.dev2
hydra-core==1.3.2
lxml==5.2.2
matplotlib==3.8.4
numpy==2.3.1
omegaconf==2.3.0
openai==1.95.1
pandas==2.3.1
Pillow==11.3.0
pyarrow==16.1.0
Requests==2.32.4
torch==2.6.0
tqdm==4.66.2
transformers==4.37.2
typer==0.16.0
